+++ src/Output/LoggerWriter.php
@@ -7,26 +7,23 @@ namespace Knotlog\Output;
 use Knotlog\Log;
 use Psr\Log\LoggerInterface;
 
-use function json_encode;
-
-use const JSON_THROW_ON_ERROR;
-
 final readonly class LoggerWriter implements LogWriter
 {
     public function __construct(
         private LoggerInterface $logger,
-        private int $flags = 0,
+        private string $messageKey = 'message',
+        private string $errorKey = 'error',
     ) {
     }
 
     public function write(Log $log): void
     {
-        $data = json_encode($log, $this->flags | JSON_THROW_ON_ERROR);
+        $context = $log->all();
 
         if ($log->hasError()) {
-            $this->logger->error($data);
+            $this->logger->error('{' . $this->errorKey . '}', $context);
         } else {
-            $this->logger->info($data);
+            $this->logger->info('{' . $this->messageKey . '}', $context);
         }
     }
 }