+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - LogResponseError sets the error to the HTTP reason phrase
+- LoggerWriter uses PSR-3 interpolation, and passes the context
 
 ## [0.1.0]